feat: support OpenCode v2 provider hooks
