+++ CHANGELOG.md
@@ -1,60 +1,5 @@
 # Changelog
 
-## Unreleased
-
-- **OpenCode v2 (beta) support** — the default export is now dual-shape:
-  `{ id, server, setup }`. OpenCode v2 validates the export against its
-  plugin schema and calls `setup(context)`; OpenCode 1.x ignores `setup`
-  and calls `server` as before (verified against a real 1.18.25 binary).
-  One package serves both generations.
-  - **v2 auto-registration** — the commandcode provider and every snapshot
-    model are injected through `catalog.transform` (the v2 catalog draft
-    upserts providers via `update` and models via the `models` map). The
-    pass is idempotent: transforms replay on every catalog rebuild, and
-    user-declared provider/model entries always win.
-  - **v2 inference routing** — the provider `package` is set to
-    `aisdk:commandcode`, which routes model resolution through
-    `aisdk.hook("language")`; the hook builds the bundled
-    `CommandCodeLanguageModel` with the runtime-resolved credential
-    (`options.apiKey`), so no npm SDK package is loaded.
-  - **v2 auth** — an integration with an `env` method
-    (`COMMANDCODE_API_KEY`) and the studio OAuth method is registered via
-    `integration.transform`; v2 resolves the credential into the language
-    hook's `options.apiKey`.
-  - **v2 reasoning** — reasoning-effort cycling is delivered through real
-    `Model.Info` `variants` (`settings.reasoningEffort`), which the runtime
-    merges into `providerOptions.commandcode.reasoningEffort`.
-  - **v2 Deals tool** — `cmd_plan_summary` is registered as a first-class
-    tool (`options: { codemode: false }`) via `tool.transform`, so its
-    results reach the model reliably on current betas.
-  - Known beta limitation: on beta-18684 the `aisdk.hook` seam can fail to
-    resolve after service restarts (`Unsupported package for
-commandcode/<model>` even with the plugin loaded and the catalog
-    populated). When the seam wires correctly, models register, the env
-    credential flows, and inference streams end-to-end (verified live).
-    Track upstream rather than working around it here.
-- **#102** — the v1 auto-registration now only sets
-  `env: ["COMMANDCODE_API_KEY"]` when the user has not declared a static
-  `options.apiKey`. On affected v1 builds an empty `{env:}` substitution
-  can block the config key (upstream opencode #34388), and a static key
-  makes the env hint pointless either way.
-- **#71** — the stateless Anthropic codec's `content_block_stop` now emits
-  `reasoning-end` for thinking blocks (with the index-derived
-  `thinking-<i>` id used by the codec's start/delta fallbacks), completing
-  the stateless reasoning lifecycle. The duplicated
-  `content_block_delta` branches were collapsed into one tri-branch
-  (text / thinking / partial_json).
-- **#72** — two stateful-parser robustness gaps closed:
-  (1) `content_block_stop` closes the thinking block with the id the
-  provider set on `content_block_start` (tracked per index), not the
-  index fallback — end id always matches start id; and
-  (2) the OpenAI parser synthesizes `reasoning-end`/`text-end` before a
-  mid-stream error, so a failed-mid-generation stream no longer strands
-  open parts (the error surfaces as an `{type: "error"}` part).
-- Registration/enrichment logic shared by both surfaces moved to
-  `src/plugin/core.ts` (`src/plugin/models.ts` is gone; no shim — import
-  from `core.js`).
-
 ## 1.6.1 - 2026-08-29
 
 Patch release. Catalog refresh to `command-code@1.38.1`, one catalog
